+++ tests/testthat/test-wrappers.R
@@ -40,7 +40,8 @@ test_that("_stat functions work", {
     )
   another_way <- iris3 %>%
     chisq_test(Sepal.Length.Group ~ Species) %>%
-    dplyr::select(statistic)
+    dplyr::select(statistic) %>% 
+    dplyr::rename(stat = statistic)
   obs_stat_way <- iris3 %>% 
     chisq_stat(Sepal.Length.Group ~ Species)
   one_more <- chisq.test(
@@ -113,6 +114,7 @@ test_that("conf_int argument works", {
   )
   
   # Check that var.equal produces different results
+  # Thanks for finding this @EllaKaye!
   set.seed(2018)
   iris_small <- iris2 %>% sample_n(10)
   no_var_equal <- iris_small %>% 
@@ -122,6 +124,18 @@ test_that("conf_int argument works", {
            var.equal = TRUE)
   expect_false(
     no_var_equal == var_equal
+  )
+  
+  shortcut_no_var_equal <- iris_small %>% 
+    specify(Petal.Width ~ Species) %>% 
+    calculate(stat = "t", order = c("versicolor", "virginica"))
+  
+  shortcut_var_equal <- iris_small %>% 
+    specify(Petal.Width ~ Species) %>% 
+    calculate(stat = "t", order = c("versicolor", "virginica"),
+              var.equal = TRUE)
+  expect_false(
+    shortcut_no_var_equal == shortcut_var_equal
   )
   
 })